feat: Add email visit summary functionality and migrate to App Router
- Add sendVisitNotesEmail edge function with Resend integration
- Add frontend API route for email functionality
- Add email buttons to dashboard and history pages
- Migrate all screens from /screens to App Router pattern
- Update getUserProgress to properly fetch booth details
- Remove unused screen files
- Fix booth query to only select existing columns
- Set up Resend API key as Supabase secret

This commit includes:
- Complete email functionality with beautiful HTML templates
- Modern Next.js App Router architecture
- Improved user experience with email summaries
- Clean codebase with removed legacy files
